// src/types/offer.ts
// generated by https://quicktype.io
export type ServerOffer = {
  id?: string;
  title: string;
  type: string;
  price: number;
  city?: ServerCity;
  location?: ServerLocation;
  isFavorite: boolean;
  isPremium: boolean;
  rating: number;
  previewImage: string;
}

export type ServerCity = {
  name: string;
  location: ServerLocation;
}

export type ServerLocation = {
  latitude: number;
  longitude: number;
  zoom: number;
}